+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Ghost Changelog
 
-_Showing 1 releases._
+### V1.1.0
+- Added Disqus Comment
+- Bumped up node modules dependencies
+- create grunt task for build